Add tests for test Utils - fromHex and toHex
This change also includes moving test extensions and test data to utils
package.
